requirements.txt
python-binance==0.7.12
pandas==1.3.4
requests==2.26.0
matplotlib==3.4.3
